moves shared to src folder
